refactor: categorized widgets instead of having one file per
