--- gui/src/qwidgets/MainWindow.py
@@ -1,57 +1,0 @@
-from PyQt5.QtWidgets import QWidget, QStackedWidget, QVBoxLayout
-from qwidgets import PedalBoardSelectWindow, BoardWindow
-from plugin_manager import PluginManager
-from modhostmanager import (
-    startModHost, connectToModHost, setUpPlugins, setUpPatch, verifyParameters
-)
-
-
-class MainWindow(QWidget):
-    def __init__(self):
-        super().__init__()
-        self.setGeometry(0, 0, 480, 800)
-
-        self.stack = QStackedWidget(self)
-        self.layout = QVBoxLayout(self)
-        self.layout.addWidget(self.stack)
-
-        # Create selection screen
-        self.start_screen = PedalBoardSelectWindow(self.launch_board)
-        self.stack.addWidget(self.start_screen)
-
-        self.board_window = None  # Placeholder for later
-
-        self.show()
-
-    def launch_board(self, selected_json):
-        """Called when a JSON file is selected to load the board"""
-        board = PluginManager()
-        board.initFromJSON(selected_json)
-        startModHost()
-        modhost = connectToModHost()
-        if modhost is None:
-            print("Failed Closing...")
-            return
-        setUpPlugins(modhost, board)
-        setUpPatch(modhost, board)
-        verifyParameters(modhost, board)
-
-        # Remove old board window if it exists
-        if self.board_window is not None:
-            self.stack.removeWidget(self.board_window)
-            self.board_window.deleteLater()
-
-        # Create new board window and add it to the stack
-        self.board_window = BoardWindow(
-            board,
-            mod_host_manager=modhost,
-            restart_callback=self.show_start_screen
-        )
-        self.stack.addWidget(self.board_window)
-        self.stack.setCurrentWidget(self.board_window)  # Switch view
-        self.board_window.setFocus()
-
-    def show_start_screen(self):
-        """Switch back to the start screen."""
-        self.stack.setCurrentWidget(self.start_screen)  # Switch back
-        self.start_screen.setFocus()
